Renamed @pro decorator into @server. @pro remains working though. (#2320)
* Renamed @pro decorator into @server. @pro remains working though.

* fixed add-in for server refactor

* resolve merge conflicts
